Refactor camera and chunk interfaces, add Light header
Refactored Camera, Chunk, ChunkManager, LayerManager, and QuadTree classes to use const-correctness and uint8_t for IDs and orders. Removed camera zoom functionality and related UI. Simplified LayerManager drawing logic. Added a new Light.hpp header. Improved DrawableObject constructors with std::move. Updated initialization and scene setup to remove zoom and improve logging. Updated binary outputs.

+++ Engine/Engine/TextureSlice/TextureSlice.hpp
@@ -11,7 +11,7 @@ namespace Engine
 {
 	struct TextureSliced
 	{
-		int id;
+		uint8_t id;
 		sf::IntRect rect;
 	};
